Initial commit of 5-3a. Created .gitignore to exclude backup filetypes.
MatLab script that parses all the data folders from the lab to extract the system input and output. The script then computes the fft of the input/output, take the highest frequency bin from the input and then computes the gain and phase delay for that bin. The theoretical frequency response is calculated for frequencies between 0.5-50Hz.